FAD-4223 refactored components because Jon had good ideas
