+++ package.json
@@ -59,10 +59,11 @@
     "redux": "^3.6.0"
   },
   "scripts": {
-    "start": "node scripts/start.js",
+    "start": "npm run fix && node scripts/start.js",
     "build": "node scripts/build.js",
-    "test": "npm run lint -- --fix && node scripts/test.js --env=jsdom",
-    "lint": "eslint \"src/**/*.js\""
+    "test": "npm run lint && node scripts/test.js --env=jsdom",
+    "lint": "eslint \"src/**/*.js\"",
+    "fix": "npm run lint -- --fix"
   },
   "jest": {
     "collectCoverageFrom": [
