+++ src/constants/text.js
@@ -1,0 +1,3 @@
+export const DKIM_INTRO_TEXT = 'Verifying DKIM is one step that can help your email get to the inbox. Not only is it an industry standard, but when configured correctly it demonstrates to ISPs that – unlike spammers – you are accountable for your email. Check DKIM, quickly with this simple DKIM Validator tool.';
+
+export const DKIM_COOKIE_NAME = '_dkimTestEmail';
